--- a/c/include/wkw.h
+++ b/c/include/wkw.h
@@ -15,7 +15,7 @@ typedef struct dataset dataset_t;
 dataset_t * dataset_open(const char * root);
 void   dataset_close(const dataset_t * handle);
 int    dataset_read(const dataset_t * handle, const uint32_t * bbox, void * data);
-int    dataset_write(const dataset_t * handle, const uint32_t * bbox, const void * data, bool is_fortran_order);
+int    dataset_write(const dataset_t * handle, const uint32_t * bbox, const void * data, bool data_in_c_order);
 void   dataset_get_header(const dataset_t * handle, struct header * header);
 void * dataset_create(const char * root, const struct header * header);
 int    file_compress(const char * src_path, const char * dst_path);