--- a/c/src/lib.rs
+++ b/c/src/lib.rs
@@ -5,7 +5,7 @@ use wkwrap as wkw;
 extern crate lazy_static;
 
 use std::ffi::{CStr, CString};
-use std::os::raw::{c_int, c_char, c_void, c_ulong};
+use std::os::raw::{c_char, c_int, c_ulong, c_void};
 
 use std::path::Path;
 use std::sync::Mutex;
@@ -19,13 +19,13 @@ pub struct Header {
     pub file_len: u8,
     pub block_type: u8,
     pub voxel_type: u8,
-    pub voxel_size: u8
+    pub voxel_size: u8,
 }
 
 fn as_log2(i: u8) -> Result<u8, &'static str> {
     match i & (i - 1) == 0 {
         true => Ok(i.trailing_zeros() as u8),
-        false => Err("Input must be a power of two")
+        false => Err("Input must be a power of two"),
     }
 }
 
@@ -38,21 +38,21 @@ fn from_header(header_ptr: *const Header) -> Result<wkw::Header, &'static str> {
         1 => wkw::BlockType::Raw,
         2 => wkw::BlockType::LZ4,
         3 => wkw::BlockType::LZ4HC,
-        _ => return Err("Block type is invalid")
+        _ => return Err("Block type is invalid"),
     };
 
     let voxel_type = match c_header.voxel_type {
-        1  => wkw::VoxelType::U8,
-        2  => wkw::VoxelType::U16,
-        3  => wkw::VoxelType::U32,
-        4  => wkw::VoxelType::U64,
-        5  => wkw::VoxelType::F32,
-        6  => wkw::VoxelType::F64,
-        7  => wkw::VoxelType::I8,
-        8  => wkw::VoxelType::I16,
-        9  => wkw::VoxelType::I32,
+        1 => wkw::VoxelType::U8,
+        2 => wkw::VoxelType::U16,
+        3 => wkw::VoxelType::U32,
+        4 => wkw::VoxelType::U64,
+        5 => wkw::VoxelType::F32,
+        6 => wkw::VoxelType::F64,
+        7 => wkw::VoxelType::I8,
+        8 => wkw::VoxelType::I16,
+        9 => wkw::VoxelType::I32,
         10 => wkw::VoxelType::I64,
-        _  => return Err("Voxel type is invalid")
+        _ => return Err("Voxel type is invalid"),
     };
 
     let block_len_log2 = as_log2(c_header.block_len)?;
@@ -66,7 +66,7 @@ fn from_header(header_ptr: *const Header) -> Result<wkw::Header, &'static str> {
         voxel_type: voxel_type,
         voxel_size: c_header.voxel_size,
         data_offset: 0,
-        jump_table: None
+        jump_table: None,
     })
 }
 
@@ -81,12 +81,12 @@ fn check_return<T>(ret: Result<T, &str>) -> c_int {
 }
 
 lazy_static! {
-    static ref LAST_ERR_MSG: Mutex<Box<CStr>> = Mutex::new(
-        CString::new("".as_bytes()).unwrap().into_boxed_c_str());
+    static ref LAST_ERR_MSG: Mutex<Box<CStr>> =
+        Mutex::new(CString::new("".as_bytes()).unwrap().into_boxed_c_str());
 }
 
 #[no_mangle]
-pub extern fn get_last_error_msg() -> *const c_char {
+pub extern "C" fn get_last_error_msg() -> *const c_char {
     LAST_ERR_MSG.lock().unwrap().as_ptr()
 }
 
@@ -96,7 +96,7 @@ fn set_last_error_msg(msg: &str) {
 }
 
 #[no_mangle]
-pub extern fn dataset_open(root_ptr: *const c_char) -> *const Dataset {
+pub extern "C" fn dataset_open(root_ptr: *const c_char) -> *const Dataset {
     assert!(!root_ptr.is_null());
     let root_str = unsafe { CStr::from_ptr(root_ptr) }.to_str().unwrap();
     let root_path = Path::new(root_str);
@@ -105,7 +105,7 @@ pub extern fn dataset_open(root_ptr: *const c_char) -> *const Dataset {
         Ok(dataset) => {
             let dataset_ptr = Box::from(dataset);
             unsafe { std::mem::transmute(dataset_ptr) }
-        },
+        }
         Err(msg) => {
             set_last_error_msg(msg);
             std::ptr::null::<Dataset>()
@@ -114,7 +114,7 @@ pub extern fn dataset_open(root_ptr: *const c_char) -> *const Dataset {
 }
 
 #[no_mangle]
-pub extern fn dataset_close(dataset_ptr: *const Dataset) {
+pub extern "C" fn dataset_close(dataset_ptr: *const Dataset) {
     assert!(!dataset_ptr.is_null());
 
     #[allow(unused_variables)]
@@ -125,7 +125,7 @@ pub extern fn dataset_close(dataset_ptr: *const Dataset) {
 }
 
 #[no_mangle]
-pub extern fn dataset_get_header(dataset_ptr: *const Dataset, header_ptr: *mut Header) {
+pub extern "C" fn dataset_get_header(dataset_ptr: *const Dataset, header_ptr: *mut Header) {
     assert!(!dataset_ptr.is_null());
     assert!(!header_ptr.is_null());
 
@@ -145,10 +145,13 @@ pub extern fn dataset_get_header(dataset_ptr: *const Dataset, header_ptr: *mut H
 }
 
 #[no_mangle]
-pub extern fn dataset_create(root_ptr: *const c_char, header_ptr: *const Header) -> *const Dataset {
+pub extern "C" fn dataset_create(
+    root_ptr: *const c_char,
+    header_ptr: *const Header,
+) -> *const Dataset {
     assert!(!header_ptr.is_null());
     assert!(!root_ptr.is_null());
-    
+
     let root_str = unsafe { CStr::from_ptr(root_ptr) }.to_str().unwrap();
     let root_path = Path::new(root_str);
 
@@ -156,7 +159,7 @@ pub extern fn dataset_create(root_ptr: *const c_char, header_ptr: *const Header)
         Ok(dataset) => {
             let dataset_ptr = Box::from(dataset);
             unsafe { std::mem::transmute(dataset_ptr) }
-        },
+        }
         Err(msg) => {
             set_last_error_msg(msg);
             std::ptr::null::<Dataset>()
@@ -165,7 +168,7 @@ pub extern fn dataset_create(root_ptr: *const c_char, header_ptr: *const Header)
 }
 
 #[no_mangle]
-pub extern fn file_compress(src_path_ptr: *const c_char, dst_path_ptr: *const c_char) -> c_int {
+pub extern "C" fn file_compress(src_path_ptr: *const c_char, dst_path_ptr: *const c_char) -> c_int {
     assert!(!src_path_ptr.is_null());
     assert!(!dst_path_ptr.is_null());
 
@@ -178,20 +181,18 @@ pub extern fn file_compress(src_path_ptr: *const c_char, dst_path_ptr: *const c_
 }
 
 fn c_bbox_to_off_and_shape(bbox_ptr: *const c_ulong) -> (wkwrap::Vec3, wkwrap::Vec3) {
-    let bbox = unsafe {
-        std::slice::from_raw_parts(bbox_ptr as *const u32, 6)
-    };
+    let bbox = unsafe { std::slice::from_raw_parts(bbox_ptr as *const u32, 6) };
 
     let off = wkwrap::Vec3 {
         x: bbox[0],
         y: bbox[1],
-        z: bbox[2]
+        z: bbox[2],
     };
 
     let shape = wkwrap::Vec3 {
         x: bbox[3] - bbox[0],
         y: bbox[4] - bbox[1],
-        z: bbox[5] - bbox[2]
+        z: bbox[5] - bbox[2],
     };
 
     (off, shape)
@@ -201,7 +202,7 @@ fn c_data_to_mat<'a>(
     dataset: &wkwrap::Dataset,
     shape: &'a wkwrap::Vec3,
     data_ptr: *const c_void,
-    is_fortran_order: bool
+    is_fortran_order: bool,
 ) -> wkwrap::Mat<'a> {
     let voxel_type = dataset.header().voxel_type;
     let voxel_size = dataset.header().voxel_size as usize;
@@ -213,10 +214,10 @@ fn c_data_to_mat<'a>(
 }
 
 #[no_mangle]
-pub extern fn dataset_read(
+pub extern "C" fn dataset_read(
     dataset_ptr: *const Dataset,
     bbox_ptr: *const c_ulong,
-    data_ptr: *mut c_void
+    data_ptr: *mut c_void,
 ) -> c_int {
     assert!(!dataset_ptr.is_null());
     assert!(!bbox_ptr.is_null());
@@ -232,17 +233,27 @@ pub extern fn dataset_read(
 }
 
 #[no_mangle]
-pub extern fn dataset_write(
+pub extern "C" fn dataset_write(
     dataset_ptr: *const Dataset,
     bbox_ptr: *const c_ulong,
     data_ptr: *const c_void,
-    is_fortran_order: bool
+    is_fortran_order: bool,
 ) -> c_int {
     assert!(!dataset_ptr.is_null());
     assert!(!bbox_ptr.is_null());
     assert!(!data_ptr.is_null());
 
     let dataset = unsafe { Box::from_raw(dataset_ptr as *mut wkwrap::Dataset) };
+
+    // dataset.write cannot cope with multichannel data if row-major order.
+    {
+        let header = dataset.header();
+        let num_channels = header.voxel_type.size() / header.voxel_size as usize;
+        if num_channels > 1 {
+            assert!(is_fortran_order);
+        }
+    }
+
     let (off, shape) = c_bbox_to_off_and_shape(bbox_ptr);
 
     let mat = c_data_to_mat(&dataset, &shape, data_ptr, is_fortran_order);
@@ -250,4 +261,3 @@ pub extern fn dataset_write(
     std::mem::forget(dataset);
     check_return(ret)
 }
-