This PR improves the write performance of row-major input data.

Before, the input data was converted by python to column-major and then passed to the rust library.
This conversion was done using `np.asfortranarray`. It turns out that this is unnecessarily slow.

In order to circumvent this issue, the python library determines whether the data is row or column-major and passes this information down to the rust lib.
The data is then converted on the fly using two buffers. 
This way, column-major order is written in any case.

The first buffer contains the data to write in the same order as the input data.
The second buffer is used only if the input data is in row-major and holds the column-major version of the first buffer.

The write performance for 1 Gibibyte on my laptop:
Before:
Write data column-major: 1.14s
Write data row-major: 46.1s

After:
Write data column-major: 1.23s
Write data row-major: 1.11s

Closes #38