@amotta 2019-10-24
These numbers are insane! Thank you very much for this pull request, Valentin!

Unfortunately, I won't have time to review before tomorrow / early next week. I hope this is okay with you.