@valentin-pinkau 2019-10-24
@philippotto thank you for your review.
I implemented your remarks and added more tests.
The row-major to column-major conversion gets much more complicated with multiple channels, as the data to be written is expected to be in channel first format. This is why this PR implements a fallback to the usage of `np.asfortranarray`  if there is more than one channel.