@valentin-pinkau 2019-10-28
@amotta Thank you for your review!
I addressed most of your remarks. I would like to postpone the multi-channel implementation to another PR.
Reading is quite fast at the moment, as numpy handles the order transparently. When it turns out to slow things down I will have a look :)